--- nodescraper/cli/daemon_cmd.py
@@ -1,52 +1,0 @@
-###############################################################################
-#
-# MIT License
-#
-# Copyright (c) 2026 Advanced Micro Devices, Inc.
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-#
-###############################################################################
-from __future__ import annotations
-
-import logging
-import sys
-
-from pydantic import ValidationError
-
-from nodescraper.redfish_events.daemon_config import load_daemon_config
-
-
-def run_daemon_cli(config_path: str, logger: logging.Logger) -> None:
-    """Load daemon config and run until interrupted."""
-    from nodescraper.redfish_events.daemon_runner import run_event_daemon
-
-    try:
-        config = load_daemon_config(config_path)
-    except FileNotFoundError as exc:
-        logger.error("%s", exc)
-        sys.exit(1)
-    except ValueError as exc:
-        logger.error("Invalid daemon config: %s", exc)
-        sys.exit(1)
-    except ValidationError as exc:
-        logger.error("Invalid daemon config: %s", exc)
-        sys.exit(1)
-    logger.info("Starting Redfish event daemon from %s", config_path)
-    run_event_daemon(config)
